restructured utils, added tox ini for tests
