+++ .isort.cfg
@@ -1,3 +1,3 @@
 [settings]
-known_third_party = cv2,fb_to_dict,flatbuffers,gRPC,google,grpc,imageio,matplotlib,numpy,open3d,pb_to_dict,pytest,quaternion,rospy,seerep,sensor_msgs,setuptools,std_msgs,tf,yaml
+known_third_party = cv2,flatbuffers,gRPC,google,grpc,imageio,matplotlib,numpy,open3d,pytest,quaternion,rospy,seerep,sensor_msgs,setuptools,std_msgs,tf,yaml
 profile =black
