+++ examples/python/gRPC/util/msg_abs/msgs.py
@@ -13,8 +13,8 @@ from seerep.fb import (
 )
 from seerep.util import fb_helper as fbh
 from seerep.util.datastructures import FrozenEnum
-from seerep.util.helper.msgs_base import MsgsFb, MsgsFunctions, expect_component
-from seerep.util.helper.service_manager import ServiceManager
+from seerep.util.msg_abs.msgs_base import MsgsFb, MsgsFunctions, expect_component
+from seerep.util.service_manager import ServiceManager
 
 C_MAX_INTEGER: int = 2147483647
