+++ examples/tests/python/gRPC/meta/test_gRPC_fb_createGeodeticCoordProject.py
@@ -3,12 +3,12 @@
 # gRPC_fb_getProjects.py
 from typing import List
 
-import fb_to_dict
 import flatbuffers
 import seerep.util.fb_helper as fb_helper
 from gRPC.meta import gRPC_fb_createGeodeticCoordProject as project_creation
 from gRPC.meta import gRPC_fb_getProjects as projects_query
 from seerep.fb import ProjectInfo as PI
+from seerep.util import fb_to_dict
 
 
 def test_gRPC_fb_createGeoProjectAndGetProjects(grpc_channel):
