+++ examples/tests/python/gRPC/pointcloud/test_gRPC_fb_sendAndQueryPointCloud.py
@@ -1,13 +1,11 @@
-from typing import List
-
-import fb_to_dict
 from gRPC.pointcloud import gRPC_fb_queryPointCloud as query_pc
 from gRPC.pointcloud import gRPC_fb_sendPointCloud as send_pc
+from seerep.util import fb_to_dict
 
 
 # todo: test query dependant on tf
 def test_sendAndQueryPCs(grpc_channel, project_setup):
-    proj_name, proj_uuid = project_setup
+    _, proj_uuid = project_setup
 
     # send pointclouds to the project
     sent_pcs = [
