+++ tox.ini
@@ -1,0 +1,26 @@
+[tox]
+requires =
+    tox >= 4
+env_list = py38, integration-examples
+
+[testenv]
+basepython = python3.8
+description = run tests
+allowlist_externals = sh
+deps =
+    pytest >= 7
+    pytest-sugar
+
+[testenv:integration-examples]
+description = run example integration tests
+deps =
+    flatbuffers
+    grpcio
+    numpy
+    protobuf
+    numpy-quaternion
+commands =
+    sh -c "/seerep/devel/bin/seerep_server_server -c/seerep/src/examples/tests/seerep.cfg > /dev/null 2>&1 &\
+            seerep_test_server_pid=$!;\
+            PYTHONPATH=/seerep/devel/include/seerep_com/fbs:$PYTHONPATH pytest examples {posargs};\
+            kill -s 2 $seerep_test_server_pid"
